use hashcash to prevent spam
also every scrap must first get the token secret (that
changes every 100 seconds) from the server before computing
the token.

the token secret is based at the current time (in intervals of 100
seconds) and a secret string that is stored at the design doc and
used by a show function to build the token secret on time.

of course the design doc is public, so is the token secret, but this
is in no way needed to the properly hashcash part of the process, as
it will require processing power and a second or two from the spammer
computer anyway. the token secret that must be fetched before each
scrap is just one more obstacle for the spammer action, but if the
spammer _really_ wants to spam, he can do it, he can fetch the secret
string previously and use it to generate its hashcashs and send a
trillion scraps.